+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/modal/ticketmodal.tsx","./src/components/ui/button.tsx","./src/components/ui/input.tsx","./src/components/ui/scroll-area.tsx","./src/components/ui/tabs.tsx","./src/components/ui/toast.tsx","./src/components/ui/toaster.tsx","./src/hooks/use-toast.ts","./src/lib/utils.ts","./src/pages/chatpage.tsx","./src/pages/loginpage.tsx","./src/redux/store.tsx","./src/redux/features/authslice.tsx"],"version":"5.6.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/modal/ticketmodal.tsx","./src/components/ui/button.tsx","./src/components/ui/input.tsx","./src/components/ui/scroll-area.tsx","./src/components/ui/tabs.tsx","./src/components/ui/toast.tsx","./src/components/ui/toaster.tsx","./src/hooks/use-toast.ts","./src/lib/utils.ts","./src/pages/chatpage.tsx","./src/pages/loginpage.tsx","./src/redux/store.tsx","./src/redux/features/authslice.tsx","./src/routing/layout.tsx","./src/routing/protectedroute.tsx"],"version":"5.6.3"}
